Exericios finmalizados de vetores

+++ Aula04/SomadorDeFracoesContinuas/Ex03.c
@@ -6,7 +6,7 @@ int main(){
     printf("Digite um numero positivo e maior que 0: ");
     scanf("%i", &num);
     for(i = 1; i < num+1; i++){
-        soma += 1/(float)i;
+        soma += 1/i;
     }
     printf("A soma e %.2f", soma);
 }
